Added Login Functionality and Started User Database
You can now put in a username and a password to log in to the system (only a simple placeholder screen has been put in). It will clear the username and password sections if a wrong combination is submitted. I also made a simple user class that holds all the information.