App will now repeatedly ping the unit
The application will now ping the unit with command UPDT, this will be used for when the user presses the physical button on the unit as well as for checking if the connection exists.

The debug button in the account selection screen has been removed, though the page still exists.